refactor: move auth.rs and auth_provider.rs into auth mod
