+++ ft-sdk/src/auth/mod.rs
@@ -1,5 +1,14 @@
+mod utils;
+
+mod db;
+
+#[cfg(feature = "auth-provider")]
+pub mod provider;
+
 #[derive(Clone)]
 pub struct UserId(pub i64);
+
+pub(crate) const SESSION_KEY: &str = "session";
 
 /// Any provider can provide any of this information about currently logged-in user,
 /// which is stored against the user in the database. The provider who drops in the
@@ -7,13 +16,8 @@ pub struct UserId(pub i64);
 #[derive(Debug, Clone, PartialEq)]
 pub enum UserData {
     VerifiedEmail(String),
-    Username(String),
-
     Name(String),
-    FirstName(String),
-    LastName(String),
     Email(String),
-    Age(u8),
     Phone(String),
     ProfilePicture(String),
     /// GitHub may use username as Identity, as user can understand their username, but have never
@@ -31,12 +35,8 @@ impl UserData {
     pub fn kind(&self) -> UserDataKind {
         match self {
             UserData::VerifiedEmail(_) => UserDataKind::VerifiedEmail,
-            UserData::Username(_) => UserDataKind::Username,
             UserData::Name(_) => UserDataKind::Name,
-            UserData::FirstName(_) => UserDataKind::FirstName,
-            UserData::LastName(_) => UserDataKind::LastName,
             UserData::Email(_) => UserDataKind::Email,
-            UserData::Age(_) => UserDataKind::Age,
             UserData::Phone(_) => UserDataKind::Phone,
             UserData::ProfilePicture(_) => UserDataKind::ProfilePicture,
             UserData::Identity(_) => UserDataKind::Identity,
@@ -106,21 +106,3 @@ pub fn session_providers() -> Vec<String> {
     todo!()
 }
 
-pub(crate) fn ud() -> Option<ft_sys::UserData> {
-    let user = ft_sys::env::var("DEBUG_LOGGED_IN".to_string());
-    match user {
-        Some(v) => {
-            let v: Vec<&str> = v.splitn(4, ' ').collect();
-            let ud = ft_sys::UserData {
-                id: v[0].parse().unwrap(),
-                username: v[1].to_string(),
-                name: v.get(3).map(|v| v.to_string()).unwrap_or_default(),
-                email: v.get(2).map(|v| v.to_string()).unwrap_or_default(),
-                verified_email: true,
-            };
-            ft_sdk::println!("Inside ud {ud:?}");
-            Some(ud)
-        }
-        None => None,
-    }
-}
